Asset.h to core_headers #280
