+++ sources/core_headers/Asset.h
@@ -1,0 +1,27 @@
+#pragma once
+
+#include <QJsonValue>
+#include <QString>
+#include <QVector>
+
+namespace kv
+{
+
+struct Asset
+{
+    QString type_name;
+    QString asset_name;
+    QString sprite;
+    QString state;
+    bool turf;
+
+    struct VariableInfo
+    {
+        QString name;
+        QString type;
+        QJsonValue value;
+    };
+    QVector<VariableInfo> variables;
+};
+
+}
